+++ crates/cdk-sql-common/src/mint/completed_operations.rs
@@ -82,11 +82,11 @@ where
         .bind("operation_id", operation.id().to_string())
         .bind("operation_kind", operation.kind().to_string())
         .bind("completed_at", operation.completed_at().unwrap_or(unix_time()) as i64)
-        .bind("total_issued", operation.total_issued().to_u64() as i64)
-        .bind("total_redeemed", operation.total_redeemed().to_u64() as i64)
-        .bind("fee_collected", operation.fee_collected().to_u64() as i64)
-        .bind("payment_amount", operation.payment_amount().map(|a| a.to_u64() as i64))
-        .bind("payment_fee", operation.payment_fee().map(|a| a.to_u64() as i64))
+        .bind("total_issued", operation.total_issued().to_u64())
+        .bind("total_redeemed", operation.total_redeemed().to_u64())
+        .bind("fee_collected", operation.fee_collected().to_u64())
+        .bind("payment_amount", operation.payment_amount().map(|a| a.to_u64()))
+        .bind("payment_fee", operation.payment_fee().map(|a| a.to_u64()))
         .bind("payment_method", operation.payment_method().map(|m| m.to_string()))
         .execute(&self.inner)
         .await?;
@@ -103,7 +103,7 @@ where
                     "#,
                 )?
                 .bind("keyset_id", keyset_id.to_string())
-                .bind("fee", fee.to_u64() as i64)
+                .bind("fee", fee.to_u64())
                 .execute(&self.inner)
                 .await?;
             }
